requirements.txt
appdirs==1.4.3
ciso8601==1.0.3
click==6.7
packaging==16.8
pyparsing==2.2.0
python-dotenv==0.6.4
requests==2.14.2
six==1.10.0
